Moving prompt to standalone mod
