--- commands/themelio-client/src/common/prompt/mod.rs
@@ -1,1 +1,0 @@
-pub mod prompt;
